Improve download strategies and add new edition
* New edition: releases we do from gitlab.com mirror of OCB. This is a
tar.gz as odoo edition download strategy.
* Adapt defaults to reflect the new download strategy
* Create a new version file that will be kept at remote. This is used to
be able to know which release was deployed both by git or by extracting
the package. In this second case, it's more important as there's no way
to recover that information.
* Use this version file to create guards. Only download and install if
the desired version (from variables) is different from the existing one
(from the version hidden file)