Redirect odoobin output (#52)
Right now, if the command fails, Ansible gets nothing in output terms (neither err), and this is bad to debug. We need to go find the log inside the host at /etc/odoo/odoo.log, but this is mixed with other log entries. A perfect approach would feed both stdout and default log.

Note: this was already fixed in tasks/main.yml